+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - web: add support for tables query parameters
 - web: add support for predefined queries parameters
 
+### Changed
+- demo: add parameters to daily avg temperature query
+
 ### Fixed
 - handle empty SQL queries